# project/openapi.yaml
openapi: 3.0.3
info:
  title: Globe Trotter API
  description: |
    Globe Trotter - Empowering Personalized Travel Planning
    
    Complete RESTful API for travel booking and itinerary planning platform.
    
    ## Authentication
    All protected endpoints require Firebase JWT token in Authorization header:
    `Authorization: Bearer <firebase-jwt-token>`
    
    ## Rate Limiting
    - Anonymous: 100 requests/hour
    - Authenticated: 1000 requests/hour
    - Premium: 5000 requests/hour
    
    ## Error Handling
    Standard HTTP status codes with consistent error response format.
  version: 1.0.0
  contact:
    name: Globe Trotter API Support
    email: api-support@globetrotter.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.globetrotter.com/v1
    description: Production server
  - url: https://staging-api.globetrotter.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Development server
security:
  - FirebaseAuth: []
paths:
...
  /destinations/{destinationId}/reviews:
    get:
      tags: [Reviews]
      summary: Get destination reviews
      operationId: getDestinationReviews
      security: []
      parameters:
        - name: destinationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: Sort reviews by
          schema:
            type: string
            enum: [newest, oldest, rating_high, rating_low, helpful]
            default: newest
      responses:
        200:
          description: Reviews retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
    post:
      tags: [Reviews]
      summary: Create destination review
      operationId: createReview
      parameters:
        - name: destinationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreateRequest'
      responses:
        201:
          description: Review created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        400:
          $ref: '#/components/responses/BadRequest'

  # Trips Management
  /trips:
    get:
      tags: [Trips]
      summary: Get user trips
      operationId: getUserTrips
      parameters:
        - name: status
          in: query
          description: Filter by trip status
          schema:
            type: string
            enum: [draft, planned, active, completed, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: User trips retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripListResponse'
    post:
      tags: [Trips]
      summary: Create new trip
      operationId: createTrip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripCreateRequest'
      responses:
        201:
          description: Trip created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        400:
          $ref: '#/components/responses/BadRequest'
  /trips/{tripId}:
    get:
      tags: [Trips]
      summary: Get trip details
      operationId: getTrip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Trip details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        404:
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Trips]
      summary: Update trip
      operationId: updateTrip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripUpdateRequest'
      responses:
        200:
          description: Trip updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        404:
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Trips]
      summary: Delete trip
      operationId: deleteTrip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Trip deleted
        404:
          $ref: '#/components/responses/NotFound'
  /trips/{tripId}/stops:
    post:
      tags: [Trips]
      summary: Add stop to trip
      operationId: addTripStop
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripStopCreateRequest'
      responses:
        201:
          description: Stop added to trip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripStopResponse'
        400:
          $ref: '#/components/responses/BadRequest'

  # Bookings Management
...
components:
  securitySchemes:
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase JWT token
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        size:
          type: integer
          description: Items per page
        total:
          type: integer
          description: Total number of items
        total_pages:
          type: integer
          description: Total number of pages

    # User schemas
...
    DestinationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        google_place_id:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        country:
          type: string
        city:
          type: string
        continent:
          type: string
        description:
          type: string
        short_description:
          type: string
        avg_rating:
          type: number
          minimum: 0
          maximum: 5
        review_count:
          type: integer
        average_price:
          type: number
        cost_of_living_index:
          type: integer
        safety_index:
          type: integer
        avg_temperature:
          type: number
        activity_categories:
          type: array
          items:
            type: string
        featured:
          type: boolean
...
    TripCreateRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        traveler_count:
          type: integer
          minimum: 1
        total_budget:
          type: number
        currency:
          type: string
          minLength: 3
          maxLength: 3
        privacy_level:
          type: string
          enum: [private, friends, public]
    TripUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        traveler_count:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [draft, planned, active, completed, cancelled]
        total_budget:
          type: number
        currency:
          type: string
        privacy_level:
          type: string
          enum: [private, friends, public]
    TripResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        traveler_count:
          type: integer
        status:
          type: string
        total_budget:
          type: number
        currency:
          type: string
        privacy_level:
          type: string
        stops:
          type: array
          items:
            $ref: '#/components/schemas/TripStopResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TripListResponse:
      type: object
      properties:
        trips:
          type: array
          items:
            $ref: '#/components/schemas/TripResponse'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    TripStopCreateRequest:
      type: object
      required: [destination_id, sequence_order]
      properties:
        destination_id:
          type: string
          format: uuid
        sequence_order:
          type: integer
          minimum: 1
        arrival_date:
          type: string
          format: date
        departure_date:
          type: string
          format: date
        notes:
          type: string
        budget:
          type: number
    TripStopResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        destination_id:
          type: string
          format: uuid
        destination:
          $ref: '#/components/schemas/DestinationResponse'
        sequence_order:
          type: integer
        arrival_date:
          type: string
          format: date
        departure_date:
          type: string
          format: date
        notes:
          type: string
        budget:
          type: number

    # Booking schemas
...
    ReviewCreateRequest:
      type: object
      required: [rating, content]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
          maxLength: 255
        content:
          type: string
        trip_id:
          type: string
          format: uuid
        activity_id:
          type: string
          format: uuid
    ReviewResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        user_name:
          type: string
        destination_id:
          type: string
          format: uuid
        rating:
          type: integer
        title:
          type: string
        content:
          type: string
        moderation_status:
          type: string
        sentiment_score:
          type: number
        helpful_votes:
          type: integer
        created_at:
          type: string
          format: date-time
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/ReviewResponse'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # Recommendation schemas
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "VALIDATION_ERROR"
            message: "Invalid request parameters"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "NOT_FOUND"
            message: "Resource not found"
...